Update README.md
Canvas of Suspense is an app that is a captivating platform dedicated to mysterious art enthusiasts. This app offers users the ability to explore a vast collection of suspenseful artworks, interact with posts through comments and shares, and even create official accounts to showcase their own creations. Designed to foster a vibrant community of dark art aficionados, Canvas of Suspense combines creativity, interaction, and discovery in a visually stunning and engaging interface.